user/requirements.txt
requests==2.31.0
PyJWT==2.4.0
python-jose==3.3.0
werkzeug==2.0.3
cryptography==3.4.7